+++ vrp-pragmatic/tests/features/format/location_custom.rs
@@ -1,13 +1,16 @@
 use crate::format::problem::*;
-use crate::format::Location;
+use crate::format::{CustomLocationType, Location};
 use crate::format_time;
 use crate::helpers::*;
 
 #[test]
-fn can_use_location_index() {
+fn can_use_unknown_location() {
     let problem = Problem {
         plan: Plan {
-            jobs: vec![create_delivery_job_with_index("job1", 0), create_delivery_job_with_index("job2", 1)],
+            jobs: vec![
+                create_delivery_job_with_order("job1", (5., 0.), 1),
+                create_delivery_job_with_order("job2", (10., 0.), 2),
+            ],
             ..create_empty_plan()
         },
         fleet: Fleet {
@@ -16,7 +19,7 @@ fn can_use_location_index() {
                     start: ShiftStart {
                         earliest: format_time(0.),
                         latest: None,
-                        location: Location::Reference { index: 2 },
+                        location: Location::Custom { r#type: CustomLocationType::Unknown },
                     },
                     ..create_default_open_vehicle_shift()
                 }],
@@ -29,8 +32,8 @@ fn can_use_location_index() {
     let matrix = Matrix {
         profile: Some("car".to_string()),
         timestamp: None,
-        travel_times: vec![0, 3, 3, 1, 0, 3, 3, 2, 0],
-        distances: vec![0, 3, 3, 1, 0, 3, 3, 2, 0],
+        travel_times: vec![0, 5, 5, 0],
+        distances: vec![0, 5, 5, 0],
         error_codes: None,
     };
 
@@ -42,21 +45,21 @@ fn can_use_location_index() {
             .tour(
                 TourBuilder::default()
                     .stops(vec![
-                        StopBuilder::default().reference(2).schedule_stamp(0., 0.).load(vec![2]).build_departure(),
+                        StopBuilder::default().custom_unknown().schedule_stamp(0., 0.).load(vec![2]).build_departure(),
                         StopBuilder::default()
-                            .reference(1)
-                            .schedule_stamp(2., 3.)
+                            .coordinate((5., 0.))
+                            .schedule_stamp(0., 1.)
                             .load(vec![1])
-                            .distance(2)
-                            .build_single("job2", "delivery"),
-                        StopBuilder::default()
-                            .reference(0)
-                            .schedule_stamp(4., 5.)
-                            .load(vec![0])
-                            .distance(3)
+                            .distance(0)
                             .build_single("job1", "delivery"),
+                        StopBuilder::default()
+                            .coordinate((10., 0.))
+                            .schedule_stamp(6., 7.)
+                            .load(vec![0])
+                            .distance(5)
+                            .build_single("job2", "delivery"),
                     ])
-                    .statistic(StatisticBuilder::default().driving(3).serving(2).build())
+                    .statistic(StatisticBuilder::default().driving(5).serving(2).build())
                     .build()
             )
             .build()