+++ vrp-pragmatic/tests/features/format/mod.rs
@@ -1,1 +1,2 @@
+mod location_custom;
 mod location_index;